rename onListening factory
